Ajuste do exemplo Spring JPA - JPA puro

+++ spring/exemplos-spring-jpa/src/main/java/br/senac/tads/dsw/exemplosspring/produto/controller/ProdutoController.java
@@ -64,7 +64,7 @@ public class ProdutoController {
     }
 
     @GetMapping("/{id}/editar")
-    public ModelAndView editar(@PathVariable("id") long id) {
+    public ModelAndView editar(@PathVariable("id") int id) {
 
         Produto prod = produtoRepository.findById(id);
         if (prod.getCategorias() != null && !prod.getCategorias().isEmpty()) {
@@ -108,7 +108,7 @@ public class ProdutoController {
     }
 
     @PostMapping("/{id}/remover")
-    public ModelAndView remover(@PathVariable("id") Long id,
+    public ModelAndView remover(@PathVariable("id") int id,
             RedirectAttributes redirectAttributes) {
         produtoRepository.deleteById(id);
         redirectAttributes.addFlashAttribute("msgSucesso",
